+++ src/plugins/index.js
@@ -1,0 +1,5 @@
+import attributePlugin from './attribute-plugin';
+import interpolatePlugin from './interpolate-plugin';
+import minimizerPlugin from './minimizer-plugin';
+
+export { attributePlugin, interpolatePlugin, minimizerPlugin };
